[oc/swift cross resolve] split ObjCExportMapper
