+++ backend.native/compiler/ir/backend.native/src/org/jetbrains/kotlin/backend/konan/objcexport/ObjCExportMapper.kt
@@ -36,7 +36,7 @@ internal abstract class ObjCExportMapper {
 
     private val methodBridgeCache = mutableMapOf<FunctionDescriptor, MethodBridge>()
 
-    fun bridgeMethod(descriptor: FunctionDescriptor) = methodBridgeCache.getOrPut(descriptor) {
+    fun bridgeMethod(descriptor: FunctionDescriptor): MethodBridge = methodBridgeCache.getOrPut(descriptor) {
         bridgeMethodImpl(descriptor)
     }
 }
@@ -108,105 +108,6 @@ internal fun ObjCExportMapper.isObjCProperty(property: PropertyDescriptor): Bool
 
 internal fun ObjCExportMapper.doesThrow(method: FunctionDescriptor): Boolean = method.allOverriddenDescriptors.any {
     it.overriddenDescriptors.isEmpty() && it.annotations.hasAnnotation(KonanBuiltIns.FqNames.throws)
-}
-
-// TODO: generalize type bridges to support such things as selectors, ignored class method receivers etc.
-
-internal sealed class TypeBridge
-internal object ReferenceBridge : TypeBridge()
-internal data class ValueTypeBridge(val objCValueType: ObjCValueType) : TypeBridge()
-
-internal sealed class MethodBridgeParameter
-
-internal sealed class MethodBridgeReceiver : MethodBridgeParameter() {
-    object Static : MethodBridgeReceiver()
-    object Factory : MethodBridgeReceiver()
-    object Instance : MethodBridgeReceiver()
-}
-
-internal object MethodBridgeSelector : MethodBridgeParameter()
-
-internal sealed class MethodBridgeValueParameter : MethodBridgeParameter() {
-    data class Mapped(val bridge: TypeBridge) : MethodBridgeValueParameter()
-    object ErrorOutParameter : MethodBridgeValueParameter()
-    data class KotlinResultOutParameter(val bridge: TypeBridge) : MethodBridgeValueParameter()
-}
-
-internal data class MethodBridge(
-        val returnBridge: ReturnValue,
-        val receiver: MethodBridgeReceiver,
-        val valueParameters: List<MethodBridgeValueParameter>
-) {
-
-    sealed class ReturnValue {
-        object Void : ReturnValue()
-        object HashCode : ReturnValue()
-        data class Mapped(val bridge: TypeBridge) : ReturnValue()
-        sealed class Instance : ReturnValue() {
-            object InitResult : Instance()
-            object FactoryResult : Instance()
-        }
-
-        sealed class WithError : ReturnValue() {
-            object Success : WithError()
-            data class RefOrNull(val successBridge: ReturnValue) : WithError()
-        }
-    }
-
-    val paramBridges: List<MethodBridgeParameter> =
-            listOf(receiver) + MethodBridgeSelector + valueParameters
-
-    // TODO: it is not exactly true in potential future cases.
-    val isInstance: Boolean get() = when (receiver) {
-        MethodBridgeReceiver.Static,
-        MethodBridgeReceiver.Factory -> false
-
-        MethodBridgeReceiver.Instance -> true
-    }
-}
-
-internal fun MethodBridge.valueParametersAssociated(
-        descriptor: FunctionDescriptor
-): List<Pair<MethodBridgeValueParameter, ParameterDescriptor?>> {
-    val kotlinParameters = descriptor.allParameters.iterator()
-    val skipFirstKotlinParameter = when (this.receiver) {
-        MethodBridgeReceiver.Static -> false
-        MethodBridgeReceiver.Factory, MethodBridgeReceiver.Instance -> true
-    }
-    if (skipFirstKotlinParameter) {
-        kotlinParameters.next()
-    }
-
-    return this.valueParameters.map {
-        when (it) {
-            is MethodBridgeValueParameter.Mapped -> it to kotlinParameters.next()
-
-            is MethodBridgeValueParameter.ErrorOutParameter,
-            is MethodBridgeValueParameter.KotlinResultOutParameter -> it to null
-        }
-    }.also { assert(!kotlinParameters.hasNext()) }
-}
-
-internal fun MethodBridge.parametersAssociated(
-        descriptor: FunctionDescriptor
-): List<Pair<MethodBridgeParameter, ParameterDescriptor?>> {
-    val kotlinParameters = descriptor.allParameters.iterator()
-
-    return this.paramBridges.map {
-        when (it) {
-            is MethodBridgeValueParameter.Mapped, MethodBridgeReceiver.Instance ->
-                it to kotlinParameters.next()
-
-            MethodBridgeReceiver.Static, MethodBridgeSelector, MethodBridgeValueParameter.ErrorOutParameter,
-            is MethodBridgeValueParameter.KotlinResultOutParameter ->
-                it to null
-
-            MethodBridgeReceiver.Factory -> {
-                kotlinParameters.next()
-                it to null
-            }
-        }
-    }.also { assert(!kotlinParameters.hasNext()) }
 }
 
 private fun ObjCExportMapper.bridgeType(kotlinType: KotlinType): TypeBridge {
@@ -305,30 +206,4 @@ internal fun ObjCExportMapper.bridgePropertyType(descriptor: PropertyDescriptor)
     assert(isBaseProperty(descriptor))
 
     return bridgeType(descriptor.type)
-}
-
-internal enum class ObjCValueType(
-        val kotlinValueType: ValueType, // It is here for simplicity.
-        val encoding: String
-) {
-
-    BOOL(ValueType.BOOLEAN, "c"),
-    CHAR(ValueType.BYTE, "c"),
-    UNSIGNED_SHORT(ValueType.CHAR, "S"),
-    SHORT(ValueType.SHORT, "s"),
-    INT(ValueType.INT, "i"),
-    LONG_LONG(ValueType.LONG, "q"),
-    FLOAT(ValueType.FLOAT, "f"),
-    DOUBLE(ValueType.DOUBLE, "d")
-
-    ;
-
-    // UNSIGNED_SHORT -> unsignedShort
-    val nsNumberName = this.name.split('_').mapIndexed { index, s ->
-        val lower = s.toLowerCase()
-        if (index > 0) lower.capitalize() else lower
-    }.joinToString("")
-
-    val nsNumberValueSelector get() = "${nsNumberName}Value"
-    val nsNumberFactorySelector get() = "numberWith${nsNumberName.capitalize()}:"
 }
